refactor(server): extract server setup into separate module
Move Fastify server configuration and setup from app.ts to server.ts to improve code organization and separation of concerns.
