+++ src/app.ts
@@ -1,40 +1,28 @@
-import Fastify, { FastifyReply, FastifyRequest } from 'fastify'
-import userRoutes from './modules/user/user.route'
-import prismaPlugin from './utils/plugins/prisma'
-import authRoutes from './modules/auth/auth.route'
-import { loggerConfig } from './utils/logger'
-import {
-    serializerCompiler,
-    validatorCompiler,
-} from 'fastify-type-provider-zod'
-import fastifyJwt from '@fastify/jwt'
+import { buildServer } from "./server";
 
-const server = Fastify({ logger: loggerConfig }).withTypeProvider();
-
-server.get(
-    '/healthceck',
-    function (request: FastifyRequest, reply: FastifyReply) {
-        reply.send({ status: 'OK' })
-    }
-)
 
 async function main() {
-    await server.register(prismaPlugin)
-    server.register(fastifyJwt, {secret: 'secret'})
-    
-    server.setValidatorCompiler(validatorCompiler);
-    server.setSerializerCompiler(serializerCompiler);
-
-    server.register(userRoutes, { prefix: 'api/users' })
-    server.register(authRoutes, { prefix: 'api/auth' })
+    const server = await buildServer();
 
     try {
-        server.listen({ port: 3000 })
+        await server.listen({ port: 3000, host: '0.0.0.0' })
         server.log.info('Server is running on port 3000')
     } catch (err) {
         server.log.error(err)
         process.exit(1)
     }
+
+    const signals = ['SIGINT', 'SIGTERM'] as const;
+    for (const signal of signals) {
+        process.on(signal, () => {
+            server.log.info(`${signal} signal received, closing HTTP server...`);
+            server.close().then(() => {
+                server.log.info('HTTP server closed');
+                process.exit(0);
+            });
+        });
+    }
 }
 
-main()
+main();
+
